pyproject.toml
[tool.poetry]
name = "pantasia-db-sync"
version = "0.1.0"
description = ""
authors = ["Erwin Wee <erwinwee.hy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pycardano = "^0.5.1"
psycopg2 = "^2.9.3"
bidict = "^0.22.0"
PyYAML = "^6.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
